requirements.txt
alembic==1.15.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpx==0.28.1
idna==3.11
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.6
pydantic==2.12.5
pydantic-settings==2.8.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.20
qrcode==8.2
razorpay==1.4.2
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
pytest==8.3.5
